Remove a few more signedness warnings

+++ src/txfonts.h
@@ -119,6 +119,6 @@ struct TexFonts : public std::vector<TexFont> {
     auto get_dimen_param(int ft, int pos) -> ScaledInt;
     void set_int_param(int ft, int pos, int v);
     void set_dimen_param(int ft, int p, ScaledInt v);
-    auto find_font(const std::string &n, int a, int s) -> int;
-    auto define_a_new_font(const std::string &n, int a, int s) -> int;
+    auto find_font(const std::string &n, int a, int s) -> size_t;
+    auto define_a_new_font(const std::string &n, int a, int s) -> size_t;
 };
